backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
SQLAlchemy==2.0.48
python-multipart==0.0.20
pydantic==2.12.5
pydantic[email]==2.12.5
qrcode[pil]==8.1
Pillow==11.2.1
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
aiosmtplib==5.1.0
email-validator==2.2.0
python-dotenv==1.1.0
httpx==0.28.1
aiofiles==24.1.0
